unity fix test v8

+++ src/include/payloads.hpp
@@ -89,7 +89,7 @@ struct location_payload
 struct heading_payload
 {
 	uint8_t     id = 2;
-	float    heading;
+	short int    heading;
 	time_stamp _last_time_stamp;
 
 	//uint8_t time_value[sizeof time_stamp];
